+++ src/pages/_document.tsx
@@ -41,6 +41,7 @@ export default class MyDocument extends Document {
             href="https://fonts.googleapis.com/css?family=Montserrat:300,400,500,700|Raleway:400,500,700&display=swap"
             rel="stylesheet"
           ></link>
+          <link rel="stylesheet" href="./animations.css" />
         </Head>
         <body>
           <Main />