+++ CHANGELOG.md
@@ -1,7 +1,22 @@
-# 0.0.0 (2020-07-31)
+#  (2020-08-13)
+
+
+### Features
+
+- Header component ([1d4c357](https://github.com/Exlabs/value-driven-templates/commit/1d4c3570202c8eef6bf1b93c251d6d8395abf106))
+- Navigation component ([498e4b9](https://github.com/Exlabs/value-driven-templates/commit/498e4b977be4cd96f56898b99094c4b90f32495c))
+- Tag component ([5c874db](https://github.com/Exlabs/value-driven-templates/commit/5c874db1a12da5f625ed77f82f4b9cb52cf112bb))
+- Template thumbnail component ([71a8f6c](https://github.com/Exlabs/value-driven-templates/commit/71a8f6cb04e88e406214bd038f6e4c4ae54e9f39))
+- Theme provider ([ccf1fbd](https://github.com/Exlabs/value-driven-templates/commit/ccf1fbd240a6c9dd7e237d13d19794a71e930b21))
+
+
+### Documentation
+
+- README update ([0afc176](https://github.com/Exlabs/value-driven-templates/commit/0afc176400094e06634b7cb2258e9e8010cc2158))
 
 
 ### Miscellaneous
 
-- Project init ([27d9ead](https://github.com/Exlabs/value-driven-templates/commit/27d9ead8bce182725e3bec657ee4da1651cffbe5))
+- Project init ([04208ec](https://github.com/Exlabs/value-driven-templates/commit/04208ec6c8d93cf27dd90b089e6ac5c784e3db96))
+- Deployment bootstrap ([6b40e51](https://github.com/Exlabs/value-driven-templates/commit/6b40e51dc284e2c72afb0ff4811623bf0fe35985))